Move yaml helpers to a separate package
So larker doesn't use node
