+++ pkg/helpers/yaml/yaml.go
@@ -1,4 +1,4 @@
-package utils
+package yaml
 
 import "gopkg.in/yaml.v3"
